Corrected angle indicator and rays
- Casted rays now are positioned correctly relative to the player view angle
- Player view angle indicator now points in correct direction

+++ Mildred.h
@@ -32,7 +32,7 @@ class Mildred {
 		static void DrawMapLines(); 
 		static SDL_Renderer* GetRenderer();
 		static void HandleUserInput();
-		static void RenderWallSlice(int* lineCollisionPointer, int drawPoint); 
+		static void RenderWallSlice(double* lineCollisionPointer, int drawPoint); 
 		static void CastRays(); 
 };
